Multiple fixes (code logic and issues)

+++ ConsoleProject/src/dto/TexHeaderMap.cs
@@ -4,8 +4,8 @@
 
         public string FileExtension { get; set; }
         public int FileVersion { get; set; }
-        public int FileCount { get; set; }
-        public int FileCountB { get; set; }
+        public short FileCount { get; set; }
+        public short FileCountB { get; set; }
         public int Unk0c { get; set; }
         public int FileListOffset { get; set; }
         public int FileInfoOffset { get; set; }
@@ -15,7 +15,7 @@
             // default constructor
         }
 
-        public TexHeaderMap(int count) {
+        public TexHeaderMap(short count) {
             this.FileExtension = "XET.";
             this.FileVersion = 1;
             this.FileCount = count;
